LayoutTests/inspector/extensions/extensions-console-expected.txt

CONSOLE MESSAGE: line 46: log message
CONSOLE MESSAGE: line 47: debug message
CONSOLE MESSAGE: line 48: warning message
CONSOLE MESSAGE: line 49: error message
CONSOLE MESSAGE: line 50: log message with two parameters, an integer: %d and a string: %s
CONSOLE MESSAGE: line 1: log message from evaluated code
CONSOLE MESSAGE: line 46: log message
CONSOLE MESSAGE: line 47: debug message
CONSOLE MESSAGE: line 48: warning message
CONSOLE MESSAGE: line 49: error message
CONSOLE MESSAGE: line 50: log message with two parameters, an integer: %d and a string: %s
CONSOLE MESSAGE: line 1: log message from evaluated code
Tests methods of webInspector.console extension API

Started extension.
Running tests...
RUNNING TEST: extension_testAddConsoleMessages
extensions-console.html:46log message console-message console-log-level
extensions-console.html:47debug message console-message console-log-level
extensions-console.html:48warning message console-message console-warning-level
extensions-console.html:49error message console-message console-error-level
extensions-console.html:50log message with two parameters, an integer: 42 and a string: foo console-message console-log-level
log message from evaluated code console-message console-log-level
debug message from extension console-message console-debug-level
test-url.html:42log message from extension console-message console-log-level
warning message from extension console-message console-warning-level
error message from extension console-message console-error-level
RUNNING TEST: extension_testGetConsoleMessages
{
    0 : {
        severity : "log"
        text : "log message"
        url : .../extensions-console.html
        line : 46
    }
    1 : {
        severity : "log"
        text : "debug message"
        url : .../extensions-console.html
        line : 47
    }
    2 : {
        severity : "warning"
        text : "warning message"
        url : .../extensions-console.html
        line : 48
    }
    3 : {
        severity : "error"
        text : "error message"
        url : .../extensions-console.html
        line : 49
    }
    4 : {
        severity : "log"
        text : "log message with two parameters, an integer: %d and a string: %s"
        url : .../extensions-console.html
        line : 50
    }
    5 : {
        severity : "log"
        text : "log message from evaluated code"
        line : 1
    }
    6 : {
        severity : "debug"
        text : "debug message from extension"
    }
    7 : {
        severity : "log"
        text : "log message from extension"
        url : test-url.html
        line : 42
    }
    8 : {
        severity : "warning"
        text : "warning message from extension"
    }
    9 : {
        severity : "error"
        text : "error message from extension"
        line : 42
    }
}
RUNNING TEST: extension_testOnConsoleMessageAdded
{
    0 : {
        severity : "log"
        text : "log message"
        url : .../extensions-console.html
        line : 46
    }
}
{
    0 : {
        severity : "log"
        text : "debug message"
        url : .../extensions-console.html
        line : 47
    }
}
{
    0 : {
        severity : "warning"
        text : "warning message"
        url : .../extensions-console.html
        line : 48
    }
}
{
    0 : {
        severity : "error"
        text : "error message"
        url : .../extensions-console.html
        line : 49
    }
}
{
    0 : {
        severity : "log"
        text : "log message with two parameters, an integer: %d and a string: %s"
        url : .../extensions-console.html
        line : 50
    }
}
{
    0 : {
        severity : "log"
        text : "log message from evaluated code"
        line : 1
    }
}
All tests done.
